requirements.txt
langchain
langgraph
openai
praw
python-dotenv
beautifulsoup4
requests
pandas
textblob
litellm
crewai
crewai_tools